Fixed binance bug where mapping amounts, changed all amount instances to float from integer

+++ src/shared/clients/events/omiga/order-book/v1/payloads_gen.go
@@ -71,7 +71,7 @@ type OrderBookEntry struct {
     
     // Money represents a Money model.
 type Money struct {
-  Amount int `json:"amount"`  // undefined
+  Amount float64 `json:"amount"`  // undefined
   Scale int `json:"scale"`  // undefined
   Currency Currency `json:"currency"` 
 }
